requirements.txt
boto3==1.36.22
botocore==1.36.22
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
urllib3==2.3.0
